+++ .circleci/setupAndTestWithFrontend.sh
@@ -52,8 +52,8 @@ cd ../project/test/frontendIntegration/
 go run main.go &
 go run main.go 8082 &
 cd ../../../supertokens-website/test/server
-npm i -d
 npm i git+https://github.com:supertokens/supertokens-node.git#$3
+npm i
 cd ../../
 npm i -d
 SUPERTOKENS_CORE_TAG=$coreTag NODE_PORT=8081 INSTALL_PATH=../supertokens-root npm test