+++ recipe/emailpassword/userContext_test.go
@@ -75,16 +75,20 @@ func TestDefaultUserContext(t *testing.T) {
 				},
 			}),
 			session.Init(&sessmodels.TypeInput{
+				GetTokenTransferMethod: func(req *http.Request, forCreateNewSession bool, userContext supertokens.UserContext) sessmodels.TokenTransferMethod {
+					return sessmodels.CookieTransferMethod
+				},
+
 				Override: &sessmodels.OverrideStruct{
 					Functions: func(originalImplementation sessmodels.RecipeInterface) sessmodels.RecipeInterface {
 						originalCreateNewSession := *originalImplementation.CreateNewSession
-						newCreateNewSession := func(res http.ResponseWriter, userID string, accessTokenPayload map[string]interface{}, sessionData map[string]interface{}, userContext supertokens.UserContext) (sessmodels.SessionContainer, error) {
+						newCreateNewSession := func(req *http.Request, res http.ResponseWriter, userID string, accessTokenPayload map[string]interface{}, sessionData map[string]interface{}, userContext supertokens.UserContext) (sessmodels.SessionContainer, error) {
 							if _default, ok := (*userContext)["_default"].(map[string]interface{}); ok {
 								if _, ok := _default["request"].(*http.Request); ok {
 									createNewSessionContextWorks = true
 								}
 							}
-							return originalCreateNewSession(res, userID, accessTokenPayload, sessionData, userContext)
+							return originalCreateNewSession(req, res, userID, accessTokenPayload, sessionData, userContext)
 						}
 						*originalImplementation.CreateNewSession = newCreateNewSession
 						return originalImplementation