+++ recipe/session/claimsWithJWT_test.go
@@ -25,13 +25,17 @@ func TestJWTShouldCreateRightAccessTokenPayloadWithClaims(t *testing.T) {
 		},
 		RecipeList: []supertokens.Recipe{
 			Init(&sessmodels.TypeInput{
+				GetTokenTransferMethod: func(req *http.Request, forCreateNewSession bool, userContext supertokens.UserContext) sessmodels.TokenTransferMethod {
+					return sessmodels.CookieTransferMethod
+				},
+
 				Jwt: &sessmodels.JWTInputConfig{
 					Enable: true,
 				},
 				Override: &sessmodels.OverrideStruct{
 					Functions: func(originalImplementation sessmodels.RecipeInterface) sessmodels.RecipeInterface {
 						oCreateNewSession := *originalImplementation.CreateNewSession
-						nCreateNewSession := func(res http.ResponseWriter, userID string, accessTokenPayload map[string]interface{}, sessionData map[string]interface{}, userContext supertokens.UserContext) (sessmodels.SessionContainer, error) {
+						nCreateNewSession := func(req *http.Request, res http.ResponseWriter, userID string, accessTokenPayload map[string]interface{}, sessionData map[string]interface{}, userContext supertokens.UserContext) (sessmodels.SessionContainer, error) {
 							if accessTokenPayload == nil {
 								accessTokenPayload = map[string]interface{}{}
 							}
@@ -40,7 +44,7 @@ func TestJWTShouldCreateRightAccessTokenPayloadWithClaims(t *testing.T) {
 							if err != nil {
 								return nil, err
 							}
-							return oCreateNewSession(res, userID, accessTokenPayload, sessionData, userContext)
+							return oCreateNewSession(req, res, userID, accessTokenPayload, sessionData, userContext)
 						}
 						*originalImplementation.CreateNewSession = nCreateNewSession
 						return originalImplementation
@@ -74,7 +78,7 @@ func TestJWTShouldCreateRightAccessTokenPayloadWithClaims(t *testing.T) {
 
 	mux.HandleFunc("/create", func(rw http.ResponseWriter, r *http.Request) {
 		var err error
-		sessionContainer, err = CreateNewSession(rw, "rope", map[string]interface{}{}, map[string]interface{}{})
+		sessionContainer, err = CreateNewSession(r, rw, "rope", map[string]interface{}{}, map[string]interface{}{})
 		assert.NoError(t, err)
 	})
 
@@ -112,6 +116,10 @@ func TestAssertClaimsWithPayloadWithJWTAndCallRightUpdateAccessTokenPayload(t *t
 		},
 		RecipeList: []supertokens.Recipe{
 			Init(&sessmodels.TypeInput{
+				GetTokenTransferMethod: func(req *http.Request, forCreateNewSession bool, userContext supertokens.UserContext) sessmodels.TokenTransferMethod {
+					return sessmodels.CookieTransferMethod
+				},
+
 				Jwt: &sessmodels.JWTInputConfig{Enable: true},
 			}),
 		},
@@ -143,7 +151,7 @@ func TestAssertClaimsWithPayloadWithJWTAndCallRightUpdateAccessTokenPayload(t *t
 
 	mux.HandleFunc("/create", func(rw http.ResponseWriter, r *http.Request) {
 		var err error
-		sessionContainer, err = CreateNewSession(rw, "rope", accessTokenPayload, map[string]interface{}{})
+		sessionContainer, err = CreateNewSession(r, rw, "rope", accessTokenPayload, map[string]interface{}{})
 		assert.NoError(t, err)
 	})
 
@@ -205,6 +213,10 @@ func TestMergeIntoAccessTokenPayloadForJWT(t *testing.T) {
 		},
 		RecipeList: []supertokens.Recipe{
 			Init(&sessmodels.TypeInput{
+				GetTokenTransferMethod: func(req *http.Request, forCreateNewSession bool, userContext supertokens.UserContext) sessmodels.TokenTransferMethod {
+					return sessmodels.CookieTransferMethod
+				},
+
 				Jwt: &sessmodels.JWTInputConfig{
 					Enable: true,
 				},
@@ -234,7 +246,7 @@ func TestMergeIntoAccessTokenPayloadForJWT(t *testing.T) {
 	mux := http.NewServeMux()
 
 	mux.HandleFunc("/create", func(rw http.ResponseWriter, r *http.Request) {
-		CreateNewSession(rw, "rope", nil, map[string]interface{}{})
+		CreateNewSession(r, rw, "rope", nil, map[string]interface{}{})
 	})
 
 	mux.HandleFunc("/verifySession", VerifySession(nil, func(rw http.ResponseWriter, r *http.Request) {
@@ -266,7 +278,7 @@ func TestMergeIntoAccessTokenPayloadForJWT(t *testing.T) {
 
 	reqV, err := http.NewRequest(http.MethodGet, testServer.URL+"/verifySession", nil)
 	assert.NoError(t, err)
-	reqV.Header.Add("Cookie", "sAccessToken="+cookieData["sAccessToken"]+";"+"sIdRefreshToken="+cookieData["sIdRefreshToken"])
+	reqV.Header.Add("Cookie", "sAccessToken="+cookieData["sAccessToken"])
 	reqV.Header.Add("anti-csrf", cookieData["antiCsrf"])
 	resv, err := http.DefaultClient.Do(reqV)
 	assert.NoError(t, err)